Tidied up code a little. The thread_event_init/destroy fucntions do not belong in the .h file as generally accessible. They are really only meaningful for the default domain which uses pthread mutex+cv to implement them. As such added API functions for it to the domain (thread_event_init/destroy).
Still to do: the waiter structure should have the thread_event be the last field as it will make
custom domains that extend thread_event easier to write. They won't have to split their state for the
thread event itself in 2 parts: one at the top with the thread_event and the rest extending the waiter_t
structure. Also, still longer term, the waiter_t itself should not dictate that the event be a thread_event.
It really just needs a base event there. Rest is really for the default domain to dictate.