Clean up handling of waiter_src.
Instead of event signal always clearing it, it is left untouched.  the owner of
the waiter has to clear it. This is better and proper and also lets lock APIs
not have to pass back a value to indicate if caller needs to wait or not.
Examining the wait_src for a non-null value can suffice. This also led to a
a couple of bugs being found and fixed.